GameAble 프로토콜 파일로 분리
