+++ RockPaperScissors/RockPaperScissors/RPSGame.swift
@@ -7,27 +7,6 @@
 
 import Foundation
 
-protocol GameAble {
-    func readyToGame()
-    func playGame() -> GameResult
-    func checkResult()
-    func resetGame()
-}
-
-protocol RPSAble: GameAble {
-    var user: Player { get }
-    var computer: Player { get }
-    
-    func updateWinCount(by result: GameResult)
-    func getGameResult() -> (winner: Player, loser: Player)
-}
-
-enum GameResult {
-    case win
-    case draw
-    case lose
-}
-
 struct RPSGame: RPSAble {
     let user: Player
     let computer: Player
